Extract common functionality into utilities.ts
